fix vectorized add kernel bug

+++ ginfer/op/kernels/vectorize.h
@@ -1,13 +1,15 @@
 #pragma once
 
+#include <cuda_fp16.h>
+
 namespace ginfer::op::kernel {
 
 template <typename T>
 struct DefaultVecSize {
-  static constexpr size_t value = 16 / sizeof(T);  // 128 bits
+  static constexpr int value = 16 / sizeof(T);  // 128 bits
 };
 
-template <typename T, size_t size>
+template <typename T, int size>
 struct alignas(sizeof(T) * size) AlignedVector {
   static_assert(size > 1 && (size & (size - 1)) == 0, 
               "size must be a power of 2 and greater than 1");
@@ -15,5 +17,6 @@ struct alignas(sizeof(T) * size) AlignedVector {
               "size exceeds the maximum vector size for type T");
   T val[size];
 };
+
 
 }  // namespace ginfer::op::kernel
